tools/runner: fixes bug where changes could be lost
fixes a bug where you may not now that a file has been changed if you
are using `--quiet`.

the code would skip notifiying you that the file has changed if using
quiet because it would supress that output, so using that and piping it
to a git add would result in those changed files not being committed and
the user not knowing that perhaps the build they are trying to commit
may not reflect the code.

now if will output an error which will cause the git add to fail and the
user can then take the approprtiate actions